Moved s3 image upload function to api/images.js
