+++ src/page-components/admin/admindoc/admindoc.json
@@ -9,7 +9,7 @@
                 },
                 {
                     "subheading": "Subheading 2",
-                    "text": "TestB"
+                    "text": "# TestB"
                 }
 
             ]
@@ -19,11 +19,11 @@
             "subheadings": [
                 {
                     "subheading": "Subheading 1",
-                    "text": "TestC"
+                    "text": "## TestC"
                 },
                 {
                     "subheading": "Subheading 2",
-                    "text": "TestD"
+                    "text": "[] TestD"
                 }
 
             ]
